feat: add heap-based queue implementation
